+++ src/lib/db/models/Users.ts
@@ -41,9 +41,9 @@ Users.init(
       allowNull: false,
     },
     deposit: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
       allowNull: false,
-      defaultValue: 0,
+      defaultValue: 0.0,
     },
     role: {
       type: DataTypes.ENUM,
